The Extension API check does not work.
Details: The Extension class constructor implementation is provided in the cpp
file. So when the extension's static object s_##name_extension is constructed,
the constructor called is in the HHVM binary and hence the m_hhvmAPIVersion
private member variable is initialized to HHVM_API_VERSION defined in the
binary itself, not what it was when the extension was compiled.
The patch moves the Extension constructor to the header file to fix the issue.